更改L序列化类型

+++ emmy_debugger/src/emmy_facade.cpp
@@ -16,7 +16,7 @@
 
 #include "emmy_debugger/emmy_facade.h"
 #include <cstdarg>
-#include <algorithm>
+#include <cstddef>
 #include "emmy_debugger/proto/socket_server_transporter.h"
 #include "emmy_debugger/proto/socket_client_transporter.h"
 #include "emmy_debugger/proto/pipeline_server_transporter.h"
@@ -614,7 +614,7 @@ void EmmyFacade::Attach(lua_State* L)
 		rapidjson::Document rspDoc;
 		rspDoc.SetObject();
 		// fix macosx compiler error
-		rspDoc.AddMember("state", reinterpret_cast<uint64_t>(L), rspDoc.GetAllocator());
+		rspDoc.AddMember("state", reinterpret_cast<ptrdiff_t>(L), rspDoc.GetAllocator());
 		this->transporter->Send(int(MessageCMD::AttachedNotify), rspDoc);
 	}
 }
